Fixed format argument type warning.

+++ tsk/fs/hfs.c
@@ -515,9 +515,9 @@ hfs_ext_find_extent_record_attr(HFS_INFO * hfs, uint32_t cnid,
                     if (nodesize < 4 || keylen + 4 > nodesize || rec_off >= nodesize - 4 - keylen) {
                         tsk_error_set_errno(TSK_ERR_FS_GENFS);
                         tsk_error_set_errstr
-                            ("hfs_ext_find_extent_record_attr: offset and keylenth of record %d in index node %d too large (%d vs %"
+                            ("hfs_ext_find_extent_record_attr: offset and keylenth of record %d in index node %d too large (%" PRIu64 " vs %"
                             PRIu16 ")", rec, cur_node,
-                            (int) rec_off + keylen, nodesize);
+                            rec_off + keylen, nodesize);
                         free(node);
                         return 1;
                     }
